third_party/blink/web_tests/platform/generic/inspector-protocol/debugger/domdebugger-set-instrumentation-breakpoint-expected.txt

Tests the DOMDebugger.setInstrumentationBreakpoint API.

Running test: testScriptFirstStatement
{
    eventName : instrumentation:scriptFirstStatement
}

Running test: testRemoveScriptFirstStatement

Running test: testInvalidEventName
{
    id : <number>
    result : {
    }
    sessionId : <string>
}
